Make the command-line mostly compatible with Sphinx3 (stuff like -ctl_lm and such still needs to be done).  Yes all the beam widths are now very different.  Also remove -phperp, as it wasn't very useful and has disappeared from Sphinx2 recently anyway (will probably merge Ravi's confidence scoring provided it's fast enough)
git-svn-id: svn+ssh://svn.code.sf.net/p/cmusphinx/code/trunk/pocketsphinx@6245 94700074-3cef-4d97-a70e-9c8c206c02f5

+++ include/search.h
@@ -109,10 +109,10 @@ typedef struct latnode_s {
 /* Interface */
 void search_initialize (void);
 void search_set_beam_width (double beam);
-void search_set_new_word_beam_width (float beam);
-void search_set_lastphone_alone_beam_width (float beam);
-void search_set_new_phone_beam_width (float beam);
-void search_set_last_phone_beam_width (float beam);
+void search_set_new_word_beam_width (double beam);
+void search_set_lastphone_alone_beam_width (double beam);
+void search_set_new_phone_beam_width (double beam);
+void search_set_last_phone_beam_width (double beam);
 
 /* Get logs2 beam, phone-exit beam, word-exit beam */
 void search_get_logbeams (int32 *b, int32 *pb, int32 *wb);
